extract Counter class into class.rb file
